part_1_OCR/requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0
azure-core==1.29.5
openai==0.28.1
numpy==1.24.3
pillow==10.1.0
pandas==2.0.3
matplotlib==3.7.2